grammar: evaluate project using grammar rules
Grammar evaluation is done for a specific set of build-on/build-for
architectures, however those are listed in the build plan which in
turn is obtained from the project model. To keep the application
focused on the project model we evaluate grammar twice, first to
obtain the build plan, and then to resolve for the correct set
of architectures.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>